Separated tests into different files to make more sense
