+++ app/controllers/questions_controller.rb
@@ -1,6 +1,6 @@
 class QuestionsController < ApplicationController
   def index
-    @questions = Question.all
+    @questions = Question.all.limit(20)
   end
 
   def show
