+++ app/models/question.rb
@@ -1,7 +1,9 @@
 class Question < ActiveRecord::Base
   belongs_to :user
-  belongs_to :accepted_answer, class_name: "Answer"
-  has_many :answers, dependent: :destroy
+  belongs_to :accepted_answer,
+    class_name: "Answer"
+  has_many :answers,
+    dependent: :destroy
 
   validates :title, presence: true, length: { in: 10..200 }
   validates :body, presence: true, length: { in: 15..10000 }
